MODE-768 Implemented the JCR2 Query Object Model interfaces (e.g., javax.jcr.query.qom) by extending the Graph API's Abstract Query Model classes. This required a number of changes to the AQM classes, including:
- Most of the getter methods were changed from the 'get<Name>()' form to just '<name>()', so that the new JCR QOM implementation subclasses can use the 'get<Name>()' form and return potentially different types than used by the AQM classes. For example, the AQM classes use a SelectorName representation (whereas the QOM uses a simple string), along with a number of enumerations (whereas the QOM uses simple strings).

- Most of the abstract classes in the AQM were transformed into interfaces. Without this, creating the QOM implementation classes would require multiple inheritance. With these changes, the QOM interface hierarchy, the QOM implementation class hierarchy, and the AQM class hierarchy all could be merged successfully.

- Because the ModeShape AQM classes are richer and have more capabilities, additional interfaces for these new components were added to the 'org.modeshape.jcr.query.qom' package in the 'modeshape-jcr-api' module, including an extension for the QueryObjectModelFactory, QueryObjectModelConstants, several new interfaces (e.g., Between, ArithmeticOperand, Limit, NodeDepth, NodePath, ReferenceValue) and several interfaces that represent a greater range of queries (e.g., QueryCommand, SelectQuery, SetQuery, and SetQueryObjectModel).

Despite all of these changes, the grammars of all supported query languages, and the string representation of the AQM (which is equivalent to our extended JCR-SQL2 grammar) remain unchanged.  Additionally, the new QOM implementation objects' string representation is simply that of the AQM (i.e., extended JCR-SQL2 grammar).

Finally, the javax.jcr.query.Query.JCR_JQOM query language was added to our JCR Repository implementation. Basically, this language's parser is a trivial subclass of our JCR-SQL2 parser. The JCR 2.0 specification does not explicitly define a textual query language or grammar for the Query Object Model (the grammar in the specification is written more in terms of the code of a Java-like language). However, per the Section 6.9 of the JCR 2.0 specification, the QueryManager.createQuery(String, String) method:

    'is used for languages that are string-based (i.e., most languages, such as JCR-SQL2) as well as for the for the string serializations of non-string-based languages (such as JCR- JQOM)'

At this point, all unit and integration tests pass.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1880 76366958-4244-0410-ad5e-bbfabb93f86b