MODE-788 The problem appears to be in the MapTransaction, which serves as one of the base classes for map-oriented connectors, including the InMemory and Infinispan connectors (among others). When a node was being removed from the transaction state, the transaction iterated over the subgraph as defined by the persistent state, so this generated an error when the deleted node was not yet persisted. The correct simply involves iterating over the subgraph within the transaction state first, and if there is none then doing so over the persistent state.
Several new tests were added to the WritableConnectorTest class that serves as a base class for various connector-specific test classes.  As such, they're automatically included in the tests for the map-oriented connectors. These tests highlighted related issues in the JBoss Cache connector and JPA connector, and these fixes are included.

And at this point, all unit and integration tests pass.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1896 76366958-4244-0410-ad5e-bbfabb93f86b