MODE-795 Created a patch that sets the auto-created properties defined on the 'mix:created' and 'mix:lastModified' mixins.
This also changes (e.g., fixes) the way autocreated properties are handled to not just check the property definitions explicitly defined on a node type, but also the implicit definitions inherited from supertypes. (The way this was called, the 'autoCreateItemsFor(...)' method on the SessionCache.NodeEditor is invoked given a primary type (when a node is created) or when a mixin is added. Therefore, the inherited autocreated properties need to be considered.)

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1902 76366958-4244-0410-ad5e-bbfabb93f86b