MODE-790 This query does seem to work correctly (note the '@' before the final 'drools:archive' criteria):
  /jcr:root/drools:repository/drools:package_area//element(*, drools:assetNodeType)[jcr:contains(., 'testQueryText*') and @drools:archive = 'false']

In this query, the '@drools:archive' clause forms an attribute test, which appears to be a more valid XPath, given that the JCR 1.0.1 specification describes (in Section 6.6.1) XPath as conceptually operating over the Document View of the repository (defined in Section 6.4.2 of JCR 1.0.1), where node properties are explicitly mapped into XML attributes (Section 6.4.2.1, #5).

Without the '@' symbol, this clause is technically mapped as a node name test that technically applies to a child element, which in the Document Model implies a child node. However, such a construct in a comparison clause does not make sense.  I would interpret this as an invalid XPath statement, because it is at odds with the Document View as described in the JCR 1.0.1 specification.

However, if one simply examines the (simplified) XPath grammar outlined in Section 6.6.6 (specifically rule 50), the '@' is optional. And because this comparison clause can only apply to attributes, this condition is actually not ambiguous: one can assume the name test is actually to be interpreted as an attribute name test.  This is the change that is outlined in the attached patch, which includes several new unit tests to cover several variations of this particular condition.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1904 76366958-4244-0410-ad5e-bbfabb93f86b