MODE-797 MODE-798 Fixes were made for both of these issues in the same commit (unfortunately).
For MODE-797: After some investigation, I discovered that the federated repository source that the JcrRepository sets up internally to project the system workspace (e.g., that containing the '/jcr:system' content) was not being configured with any of the workspaces defined in the source. When a session was created, the logic did check that the workspace did exist in the internal federated repository source, and if not the workspace was added to the federated repository source. This means that one could always create a JCR session for any of the underlying source's workspaces. But it mean that Workspace.getAvailableWorkspaceNames() only returned the names of the workspaces for which a session had been established.

This problem was corrected by altering the logic of how the internal federated repository source was initially configured in the JcrRepository constructor, so that a federated workspace was created for each of the workspaces in the underlying source. Note that if a workspace is created in the underlying source _after_ the JcrRepository was instantiated, such workspaces will not be included in the list of available workspace names until a session is created to that new workspace.

For MODE-798: Added support for these methods, added new ModeShapePrivileges for creating and deleting workspaces, and properly configured the Repository.OPTION_WORKSPACE_MANAGEMENT based upon source capabilities (unless the descriptor map passed into the constructor is already configured with a value of 'false'). Notice with the latter that it is possible to override this descriptor by setting it to false, ensuring that no user of the repository has the ability to create or delete workspaces (even if their permissions enable it). Also updated the Reference Guide to reflect the new ModeShapePrivileges and ModeShapeRoles.ADMIN behavior.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1905 76366958-4244-0410-ad5e-bbfabb93f86b