MODE-802 The o.m.jcr.api.QueryResult.getColumnTypes() method implementation (in o.m.jcr.query.JcrQueryResult and its subclasses) had been using the column names to resolve to the schemata. However, with the previous change, the column names contained the aliases (as expected), and there was no way to resolve the aliases to the correct table in the schemata.  The JDBC driver integration tests do verify the type names, and therefore failed.
I've been wondering for the last couple days whether it would be useful to have the column types on the PROJECT nodes. There are a couple of advantages to this:
1) The query planning/optimization/validation process is already processing the query plan and using the schemata, and it has access to the aliases and the tables they resolve to. Seems a bit odd for other components to use parts of the query plan but then require them to go back to the schemata for type information.
2) Recording the type information in the plan would mean other components that need the type information can access it directly.
3) The query plan would show more explicitly the types for the various projects, and this is helpful in debugging a query.

For these reasons I decided to add support for column types in several key places, including the QueryResults.Columns interface. Doing this required a number of constructors and signatures to change, affecting quite a few test cases. However, I believe the changes are actually for the better and are noticeably more cohesive.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1918 76366958-4244-0410-ad5e-bbfabb93f86b