MODE-807 Provide a RepositoryProvider Implementation that Uses RepositoryFactory
Applied a first version of this patch that adds the FactoryRepositoryProvider as a new RepositoryProvider that relies on JcrRepositoryFactory to access JcrEngines and JcrRepository instances.

This patch doesn't remove the dependency from modeshape-web-jcr to modeshape-jcr though, so there's still some work to be done.  In particular, the o.m.j.api.RepositoryFactory interface will probably need to be expanded to add some methods that are currently in JcrRepositoryFactory and used by FactoryRepositoryProvider.  Then FactoryRepositoryProvider could be changed to use the o.m.j.api.RepositoryFactory interface.  Also, the documentation (e.g., Getting Started Guide, Reference Guide) needs to be updated to use FactoryRepositoryProvider as the new recommended implementation of RepositoryProvider.

Nonetheless, the current version is usable and the REST and WebDAV integration tests have been changed to use FactoryRepositoryProvider.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1971 76366958-4244-0410-ad5e-bbfabb93f86b