MODE-796 Changed the behavior of the GraphSession to track node locations as they are moved, using the actual location of the requests after they are executed. Additionally, the JCR Node implementations were changed to have their location be updated whenever the node is moved (before the save as well as being found after the save).
Incidentally, all of the Request objects were not allowing actual locations to be different than the input locations. This is obviously a problem if a batch contains a move and a read of the moved node; the move will change the location of the node, while the second read request will still have the old location as the input and the output will be different than the input location. All of the Request classes were updated to remove these assertion and argument checks.

A new test case was added to verify the new behavior of JCR nodes after being moved. At this point, all unit and integration tests pass.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1972 76366958-4244-0410-ad5e-bbfabb93f86b