MODE-786 Changed the JcrRepository and JcrObservationManager to generate events for changes in the underlying system storage. Several test cases were added to verify that changes under '/jcr:system' for version history, namespaces, and locks were all being represented in events. Additionally, the JCR TCK unit test that was commented out now works.
Incidentally, this uncovered a bug in the base connector implementation. Transactions were being created with the RepositorySource's ExecutionContext rather than the ExecutionContext passed in to the RepositoryConnection.execute(...) method. This didn't cause a problem, except when user data was set via the ObservationManager interface; the user data was captured on the session's ExecutionContext (the one passed into the RepositoryConnection.execute(...) method), but was obviously not captured on the RepositorySource's ExecutionContext. The fix was simply to pass the correct ExecutionContext reference into the appropriate methods.

All unit and integration tests pass.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1976 76366958-4244-0410-ad5e-bbfabb93f86b