MODE-677 Completed the initial implementation of the clustering functionality, which at this point appears to be working as expected. These changes add test cases and fixes to the wiring of the JcrEngine, ModeShapeEngine, RepositoryService, and RepositoryLibrary with respect to the observation busses. The RepositoryLibrary was instantiating its own observation bus, so this patch requires passing it in via the RepositoryLibrary and RepositoryService constructor.
Another change that was made was how the JcrRepository's observer monitors the changes coming from the ClusteringService. The JcrRepository and its JcrSession objects make local changes via the federated repository source (and not via the actual repository source in the RepositoryLibrary) and the JcrRepository's observer was looking only for Changes coming from that federated source. However, because the federated source is not in the RepositoryLibrary, its Changes are not propagated via the ClusteringService, so the JcrRepository's observer would only ever see its own local changes. This patch changes this so that the JcrRepository's observer also watches for Changes on the actual repository source (in addition to those local changes on the federated source) that were made by other engine instances, and these are adapted into a new Changes object that looks like it came from the federated source. The end result is that the Session's see all changes correctly, whether they originated locally in the same JcrEngine or remotely in other JcrEngines.

All unit and integration tests pass.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1992 76366958-4244-0410-ad5e-bbfabb93f86b