MODE-807 Provide a RepositoryProvider Implementation that Uses RepositoryFactory
Applied patch that adds a new SecurityContext class at the modeshape-jcr-api level.  This new class provides the same functionality as the SecurityContext in modeshape-graph, but allows classes that only depend on modeshape-jcr-api (like the web classes) to function without having to pull in modeshape-jcr or modeshape-graph and all of their dependencies.

The FactoryRepositoryProvider was changed to use ServiceLoader to load  the modeshape-jcr-api RepositoryFactory in a manner comparable to how JCR 2.0 RepositoryFactory instances are intended to be loaded.

All tests pass.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@1999 76366958-4244-0410-ad5e-bbfabb93f86b