MODE-820 Changed the DDL parsers to put the effective length (which includes the supplied value and multiplied by the 'K', 'M', or 'G' factor) into the AST.
The DataType class already had fields to track the KMG factor, but this was not used consistently. After a bit of discussion, it was decided that these fields should be removed and the effective length value should be placed into the DataType object. And, because this length value can now be quite large, it was changed from an int type to a long type.

At this point, all unit and integration tests pass.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@2006 76366958-4244-0410-ad5e-bbfabb93f86b