MODE-833 Added support for '*' and '%' as zero-or-one character wildcards, and '?' and '_' as single-character wildcards. All of these can be used as literals in the terms only when they are escaped with the '\' character (which itself can be escaped when needed as a literal character).
These wildcards are available for use in the second parameters of the 'jcr:contains(...)' expressions in XPath queries and the 'CONTAINS(...)' expressions in JCR-SQL and JCR-SQL2 queries. These wildcards are also available for use in the full-text search language. New unit tests were added in several different modules.

The Reference Guide was updated to document this new capability. All unit and integration tests pass.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@2062 76366958-4244-0410-ad5e-bbfabb93f86b