MODE-834 The XPath-to-JCRSQL2 translator was not inserting any special pattern for any path segment missing a literal SNS index value. And thus the XPath behavior was identical to the JCR-SQL and JCR-SQL2 behavior. However, this was incorrect.
Therefore, the fix is to change the translator to insert '[%]' into the JCR-SQL2 PATH(...) expression for each path segment missing an SNS index. This way the JCR-SQL and JCR-SQL2 behaviors remain the same, while the XPath behavior is corrected. Several new test cases were added to verify this new XPath behavior. All unit and integration tests pass. And the Reference Guide was updated to correct some of the XPath examples and to correctly describe the behavior.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@2067 76366958-4244-0410-ad5e-bbfabb93f86b