MODE-837 Corrected several CND files that used the older MULTIPLE keywords, and also changed the CND import flag to 'true' to allow pre-JCR 2.0 CND formats. All unit and integration tests run.
I also created an integration test that configures an engine with all of the sequencer and connector CND files, and in the process I noticed that it was difficult to specify a lot of (longer) CND file paths and keep the XML readable. So I changed JcrEngine to treat the 'mode:resources' property on 'jcr:nodeTypes',  as a multi-valued property (where each value can contain a comma-delimited list of CND paths).

he Reference Guide was updated to reflect this. Again, all unit and integration tests pass.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@2071 76366958-4244-0410-ad5e-bbfabb93f86b