MODE-810 RepositoryNodeTypeManager should respond when stored node types are changed (in clustered environment)
Applied a patch that makes RepositoryNodeTypeManager implement JcrSystemObserver and be registered in JcrRepository.  This allows RTNM to listen for changes to the registered node types in the system graph and update itself accordingly.  The current implementation is fairly brutish in that any change to the projected node types causes other nodes in the cluster to reload their entire node type hierarchy.  A separate defect will be opened to optimize this in the future.  Test cases have been added for node type registration and unregistration in a cluster.

The patch makes a number of ancillary changes.  First, the default value of the PROJECT_NODE_TYPES option in JcrRepository was changed from false to true.  The node type clustering piece works by listening for changes to that projection.  Any repository that does not project node types will not work correctly in a cluster.  Changing this value did not seem to have an appreciable impact on the total execution time for the TCK test on my local machine.

A bug in JcrNodeTypeManager.createNodeTypeTemplate(NodeTypeDefinition) was identified and corrected.  JNTM was previously not properly setting values on any child nodes in the new node type template.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@2077 76366958-4244-0410-ad5e-bbfabb93f86b