MODE-839 Make RepositoryNodeTypeManager only reload changed nodes during cluster update
Applied a patch that adds a new argument to GraphNodeTypeReader.read(Graph, Path) that, if not null, restricts the GNTR to only reading node types with the provided names.  The patch then changes RepositoryNodeTypeManager to collect the names of created and deleted nodes from the Changes argument in the notify(Changes) method.  The RTNM then only removes the deleted nodes and only re-adds the newly created nodes.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@2109 76366958-4244-0410-ad5e-bbfabb93f86b