MODE-31 Changed the model sequencer to set default property values if they're missing. This is because the search engine indexing logic uses the graph layer, so the graph layer needs to have the default values for properties (otherwise these properties will have no values). The default values are read off of the 'teiid.cnd' file using a custom utility.
Also added several integration tests to verify the 'teiid.cnd' file (which is merely an aggregation of the other CND files). This led to the discovery of several problems with the node type definitions, including the fact that node types can only define children with expected primary types, so in order to constraint node types on children we need to make our node types non-mixins.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@2185 76366958-4244-0410-ad5e-bbfabb93f86b