MODE-31 MODE-680 Completed the initial implementation of the Teiid model and VDB sequencers. Additional testing needs to be completed, but at this point both sequencers appear to be operating very well against a variety of test files.
The ModelSequencer can be used to extract the structural information contained with a Teiid relational XMI model, including the catalogs, schemas, tables, views, columns, primary keys, foreign keys, indexes, procedures, procedure paramters, procedure results, logical relationships, and the JDBC source from which the model was imported. A series of node types and mixin types were defined to prescribe this structure, though all are extremely adaptable and tolerant of additional properties (as all non-mixin node types extend nt:unstructured). References within the model being sequenced are always resolved (resulting in the WEAKREFERENCE properties being set), although references to objects in other models are captured with the hrefs and/or UUIDs as defined in the model, allowing them to be resolved at another time but still enabling them to be used in queries. Descriptions, tags, and transformation mappings are all projected onto the main relational object, making it much easier to access, query, and use this information. The SHA-1 hash of the XMI file is also captured.

The VdbSequencer is used to extract the structural information contained within the Teiid relational XMI models embedded within the VDB. Although VDB files use the ZIP archive format, VDB archives contain a number of other files (e.g., runtime indexes) that are not needed, and thus the VdbSequencer does **not** simply extract the model files. Instead, when the VdbSequencer sequences a VDB archive, the virtual database's manifest is processed to identify the set of models (along with information used by the Teiid server, such as the JNDI name of the source used by the connector, the translator name, etc.), the model dependencies, and extra metadata (e.g., descriptions, problem markers, etc.). The VdbSequencer then sequences each of the embedded relational XMI models to extract the relational schema defined in the models. The output of the sequencer, then, is a single subgraph that has:
 - a root node representing the virtual database (named according to the VDB archive, minus the version information and the .vdb extension), with properties for the SHA-1 hash of the VDB archive file, version information, description, and other manifest information;
 - the placed under the VDB node the sequenced output for each relational model, including any error/warning/info problem markers that were in the VDB.

Note that the VdbSequencer resolves all (valid) references between all models within the VDB, meaning the sequenced representations of all models within the VDBs are are consistent, immediately usable, and independent of all other content in the repository (outside of the VDB subgraph).

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@2188 76366958-4244-0410-ad5e-bbfabb93f86b