MODE-869 Added support for (non-correlated) subqueries inside the WHERE clause of JCR-SQL and JCR-SQL2 queries, wherever a static operand can be used. It is even possible for subqueries to be used within other subqueries.
All subqueries must return a single column, and each row's single value will be treated as a literal value. If the subquery is used as the static operand in a clause that expects a single value (e.g., in a comparison), only the first value in the subquery's first row will be used. If the subquery is used as the static operand in a clause that allows multiple values (e.g., within a set criteria, or 'IN' clause), then the first value in each of the subquery's rows will be used. For example, this query:

  WHERE ... [my:type].[prop1] IN ( SELECT [my:prop2] FROM [my:type2] WHERE [my:prop3] &lt; '1000' ) AND ...

will use the results of the subquery as the literal values in the IN clause.

Additional unit tests for several components were created to verify behavior, and new tests in JcrQueryManagerTest were created to verify that the subqueries could be executed correctly.

The Reference Guide's section on the JCR-SQL and JCR-SQL2 query languages was updated to reflect support for subqueries and to outline the grammar for subqueries.

At this point, all unit and integration tests are successful.

git-svn-id: https://svn.jboss.org/repos/modeshape/branches/2.2.x@2227 76366958-4244-0410-ad5e-bbfabb93f86b