MODE-883 When the shareable node is removed from it's location and it has been shared to at least one other location, ModeShape moves the node from it's original location in place of the first shared location (e.g., the first proxy). However, in order to be more efficient, this move logic moved the node before removing the proxy. So for a moment, there are two nodes (the shareable node and its proxy) under one parent, and this results in an error if the parent does not allow same-name-siblings (or just one child node).
The correction involved changing the logic to first remove the proxy, then move the node to the parent, and finally to order the shareable node into the same location as the (previously removed) proxy. To reorder, we need to know the node immediately following the proxy (before removing it), so a new helper method was added to SessionCache.Node.

A new test case was added to ShareableNodeTest that replicated the test case described above (though this new test case is a bit simpler), and the problem was replicated before the fix was applied. After the fix, the test case passes, as do all other unit and integration tests.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@2239 76366958-4244-0410-ad5e-bbfabb93f86b