MODE-892 Changed the CndNodeTypeReader and the JackrabbitXmlNodeTypeReader methods to try more things when resolving files given a string path. The code used to just look on the classpath, and even then just using getClass().getResourceAsStream() method (meaning absolute paths with a leading '/' were treated as absolute paths, but relative paths without a leading '/' were considered relative to the 'org.modeshape.jcr' package).
These methods now attempt to resolve the string path into an InputStream as follows:
1) Using the ClassLoader returned from the ExecutionContext.getClassLoader(), attempt to resolve using ClassLoader.getResourceAsStream(path); or
2) Using the reader's getClass() method, attempt to resolve using Class.getResourceAsStream(path); or
3) Using the ClassLoader from the reader's Class, attempt to resolve using ClassLoader.getResourceAsStream(path); or
4) Create a java.io.File with the supplied (absolute or relative) path, and return a FileInputStream if the file exists and is readable; or
5) Create a java.net.URL with the supplied path, and (if not malformed) return the stream resulting from URL.openStream(); or
6) throw an IOException

The first InputStream returned from these methods is then passed to the reader's 'read(InputStream)' method.  This logic is actually encapsulated in IoUtil.getResourceAsStream(String).

Several additional unit tests were added for the CndNodeTypeReader and the JackrabbitXmlNodeTypeReader. Some of these verified that files could be found using the various path formats, while others verified that an IOException is now thrown if none of these resolve to a valid InputStream.

Additionally, JcrEngine was changed to use the same IoUtil.getResourceAsStream() to load the CND files referenced in the configuration files.
Interestingly, this change uncovered an error in the ModeShape repository configuration used in the unit tests in 'modeshape-jdbc', as that configuration was specifying CND files that did not exist. (Likely, this was a copy-and-paste error.) This configuration file was corrected.

The Reference Guide was also updated to describe the additional ways in which ModeShape now resolves CND files referenced in the configuration file.

All unit and integration tests pass with these changes, and they were committed into SVN on 'trunk' in r2255 and on the '2.2.x' branch in r2256.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@2255 76366958-4244-0410-ad5e-bbfabb93f86b