MODE-878 One of the major issues was that the REST client library was not properly representing the node types and type inheritance, making it difficult (if not impossible) for the JDBC metadata to properly create the table and column metadata. Also, the mapping of nodes w/ properties to tables w/ columns was slightly incorrect in its handling of residual properties and multi-valued properties (neither of which should not be mapped into columns). This filtering was done in the REST client, but should have instead been done inside the JDBC driver.
The attached patch fixes the REST client's handling node types, adopting the JCR interfaces for node types, property definitions, and child node definitions. This has a couple of advantages: the REST client doesn't need to duplicate the non-trivial interfaces and semantics; and the JDBC driver uses the REST client for remote access and the JCR interfaces for local access, and the same node type interfaces are used in both and fairly significantly simplify the JDBC metadata logic. The only disadvantage is that the JCR API library is now a dependency of the REST client, but it's small and has no other dependencies, so this is not really that bad of a change. The benefits greatly outweigh the disadvantages.

The two methods in the IRestClient interface that deal with node types also didn't really work, and we could not maintain backward compatibility because of a change in the return type. (The existing methods could have been left, but could not have been implemented correctly without duplicating a lot of code.) Therefore, the 'getNodeTypes(...)' method signature was changed to work with the usage of the JCR interfaces and the removal of unnecessary parameters.

Additionally, the 'getNodeType(...)' method that returned a single node type could not functionally work without a lot of effort and potentially numerous remote calls. Because of this, and because the return type needed to be changed and would result in no benefit over the 'getNodeTypes(...)' method, this method was simply removed.

The attached patch also changes the JDBC driver to use the new REST client's 'getNodeTypes' method and the JCR node type interfaces. The former change corrected the fundamental problem reported by this defect, while the latter allowed the REST-oriented mechanism to directly expose the JCR node type interfaces (just like the other mechanisms) and allowed for removal of a lot of code.

The result appears to be that the REST client is now returning accurate and complete representations of the node types, and the JDBC driver is now returning accurate database metadata descriptions of the available tables and columns.

git-svn-id: https://svn.jboss.org/repos/modeshape/branches/2.2.x@2272 76366958-4244-0410-ad5e-bbfabb93f86b