MODE-878 Ahem, the unit tests fail on Hudson (which is in a different time zone) due to time conversion problems. The conversion from PropertyType.DATE types to long, double, and BigDecimal values did not first convert the date into UTC.
Stupid java.util.Calendar and java.util.Date! The latter does not have the notion of a time zone, so SimpleDateFormat.parse(String) always converts the date to the current time zone, which messes everything up.

After quite a bit of time spent trying to get Java's Date and Calendar classes to work, I punted and added a dependency on JodaTime. The library is not really that big, plus the parsing and conversion logic is now the same as what's used in our JCR implementation. I also manually verified the expected milliseconds used in the unit tests, and they appear to be correct (and in UTC). The unit and integration tests pass locally, and hopefully they'll pass on Hudson.

git-svn-id: https://svn.jboss.org/repos/modeshape/branches/2.2.x@2275 76366958-4244-0410-ad5e-bbfabb93f86b