MODE-906, MODE-907 Attempted to diagnose the latency issues with the file system connector when accessing a remote file system, and found several things.
First, when the repository starts out and has empty Lucene indexes, ModeShape scans the full repository content to populate the indexes, and this scanning was very heavy-handed and could very well have attempted to load all file content into memory. Obviously that's way bad. The changes include adding a JcrRepository.Option called 'INDEX_READ_DEPTH' that will control the maximum depth of the subgraph read requests issued by the indexing process. (This works very similarly to the existing JcrRepository.Option called 'READ_DEPTH'.) Plus, the file system connector and SVN connectors now limit the maximum depth of the subgraphs they'll read to 2 (even if the ReadSubgraphRequest has a depth greater than this). The effect should be a reduction in the amount of information held in-memory at any one time during indexing.

Second, the file system connector is reading way more information than it needs. For example, when materializing an 'nt:folder' node, the connector reads the corresponding java.io.File object, but to populate the list of children, it loads the child nodes (lets say some 'nt:file' nodes), but to load those nodes, the connector needs their children (e.g., 'nt:resource' nodes, which have binary content). The problem is that the PathTransaction class is not overriding several methods, resulting in this extremely inefficient algorithm. This was corrected, so that the various Transaction methods are loading the children more efficiently, as well as computing the paths most efficiently.

Third, when the 'nt:resource' nodes are read in, the 'jcr:data' BINARY property is populated with the content, and at the moment this is done with an in-memory Binary implementation. Thus when a lot of large files are materialized, there is a lot of content being held in-memory. This, I think, is the crux of MODE-907.  This was fixed by using a new Binary implementation that will load the contents only when the Binary.getStream() method is called.

Fourth, trace logging messages were added in the file system connector's method that materializes nodes, outputting the time required to materialize each node (along with the node's type and path). This was used to verify the second finding above, and it can be used in deployments should there be any additional latency with the FS connector.

Several new test cases were added, and a few existing test cases were modified.

Also made a few changes to the FileSystem and SVN connectors to exclude all children of 'nt:file' nodes from the results of reading subgraphs (unless the 'nt:file' is the top of the subgraph). This attempts to prevent loading multiple 'nt:resource' nodes (and the potentially expensive 'jcr:data' property) inside a subgraph.

All unit and integration tests pass with these changes.

git-svn-id: https://svn.jboss.org/repos/modeshape/branches/2.2.x@2282 76366958-4244-0410-ad5e-bbfabb93f86b