MODE-911 Found a problem in the way the REST client's NodeType implementation is always including 'nt:base' as a supertype if there are no other declared supertypes. This is true for all non-mixin node types, but it is definitely incorrect for mixin node types.
This bug would result in the JDBC driver incorrectly reporting 'jcr:primaryType' as a column on tables representing mixin node types (like 'mix:created'), and when this metadata is used in a Teiid model, the resulting 'SELECT * ...' query would be expanded by Teiid into a SELECT with the unavailable table.

So I think this simple bug explains all the behavior. A simple fix to the REST Client's NodeType implementation should work.

All unit and integration tests pass.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@2298 76366958-4244-0410-ad5e-bbfabb93f86b