MODE-903 Indeed LinkedList.get(int) does not perform well with large sizes. Therefore, the suggestion to change from a LinkedList to an ArrayList because of random access (e.g., 'get(int)') is a good one. Because these arrays are created in the MapNode's 'withChild*(...)' methods, the code will only be called when the node's children are being changed. IOW, it should be backward compatible with prior serialized versions.
Since this will only kick in when nodes are changed, we still need to handle deserialized instances that have large numbers of children. To do this, the MapTransaction.Children constructor can check the type of the array of UUIDs, and copy the list using an ArrayList if the list is not an java.util.AbstractList. Note this fix does not copy the list if the list is small enough that random access will not suffer. (Obviously this is a tradeoff.)

Attached is a patch that implements this change. All unit and integration tests pass.

git-svn-id: https://svn.jboss.org/repos/modeshape/trunk@2300 76366958-4244-0410-ad5e-bbfabb93f86b